fix typing & split files
Fixed return types for `get_stress` and `get_tangent` methods.
Split constitutive laws into separate files.
