+++ structuralcodes/core/base.py
@@ -87,13 +87,13 @@ class ConstitutiveLaw(abc.ABC):
         cls.constitutive_law_counter += 1
 
     @abc.abstractmethod
-    def get_stress(self, eps: ArrayLike) -> ArrayLike:
+    def get_stress(self, eps: ArrayLike) -> t.Union[float, ArrayLike]:
         """Each constitutive law should provide a method to return the
         stress given the strain level.
         """
 
     @abc.abstractmethod
-    def get_tangent(self, eps: ArrayLike) -> ArrayLike:
+    def get_tangent(self, eps: ArrayLike) -> t.Union[float, ArrayLike]:
         """Each constitutive law should provide a method to return the
         tangent at a given strain level.
         """
@@ -104,7 +104,9 @@ class ConstitutiveLaw(abc.ABC):
         ultimate strain (negative and positive).
         """
 
-    def preprocess_strains_with_limits(self, eps: ArrayLike) -> ArrayLike:
+    def preprocess_strains_with_limits(
+        self, eps: ArrayLike
+    ) -> t.Union[float, ArrayLike]:
         """Preprocess strain arrays setting those strains sufficiently
         near to ultimate strain limits to exactly ultimate strain limit.
         """
